Attribute Macro gd_rehearse_macros::gdbench

#[gdbench]

Benchmark for gdext classes and functions integrated with Godot.

This macro transforms and registers the annotated function for further usage by GdTestRunner within a Godot test scene. When the runner enters the scene, it executes all qualified benchmarks.

A function annotated with #[gdbench] must:

Have a return value.

Have no parameters or only a singular CaseContext.

Every benchmark is executed 200 times for a warm-up, followed by 501 additional runs to assess runtime (an odd number of runs for easy median extraction). Minimum and median run times will be displayed.

Attributes

An attribute-less macro will make the benchmark run, but several attributes are available for better customizability, especially when working on specific attributes and creating more narrowly-focused test runner scenes.

skip: Skips the benchmark during execution.

focus: Forces a focused run, in which only benchmarks annotated with focus will be executed.

keyword: A specific keyword that will be picked up by the runner. The benchmark runs only if the runner has the same keyword specified.

repeat: Specifies the number of internal repeats the benchmark should undergo. By default, the function executes 100 times within every run.

Examples

use gd_rehearse::CaseContext;
use gd_rehearse::bench::*;
use godot::obj::InstanceId;

// Causes a focus run during which only the focused benchmarks will be executed, but only with
// `my bench` as a keyword in the runner.
#[gdbench(focus, keyword = "my bench")]
fn focused_bench() -> i32 {
    1337
}

// Will be skipped during the default run.
#[gdbench(skip)]
fn skipped_test() -> i32 {
    231 + 312
}

// Can access the `GdTestRunner` scene_tree.
#[gdbench]
fn bench_with_ctx(ctx: &CaseContext) -> InstanceId {
    ctx.scene_tree.instance_id()
}